backend/pyproject.toml
# SPDX-FileCopyrightText: Magenta ApS <https://magenta.dk>
# SPDX-License-Identifier: AGPL-3.0-only
[project]
name = "OpenStream"
version = "0.1.0"
description = "OpenStream"
authors = [
    {name = "Magenta ApS",email = "info@magenta.dk"}
]
license = "AGPL-3.0"
readme = "README.md"
homepage = "https://openstream.dk/"
repository = "https://git.magenta.dk/os2borgerpc/openstream/openstreamadminsite"
keywords = ["openstream"]
packages = [{ include = "openstream_django" }]

requires-python = ">=3.12,<4"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asgiref==3.9.1",
    "attrs==24.3.0",
    "boto3==1.28.54",
    "botocore==1.31.85",
    "certifi==2024.12.14",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "cssbeautifier==1.15.4",
    "defusedxml==0.7.1",
    "distlib==0.3.9",
    "dj-database-url==2.3.0",
    "Django==5.2.5",
    "django-browser-reload==1.18.0",
    "django-cors-headers==3.14.0",
    "django-environ==0.9.0",
    "django-storages==1.14.6",
    "django_csp==3.8",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "EditorConfig==0.17.1",
    "et_xmlfile==2.0.0",
    "feedparser==6.0.11",
    "filelock==3.18.0",
    "gunicorn==22.0.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "identify==2.6.13",
    "idna==3.10",
    "jmespath==1.0.1",
    "jsbeautifier==1.15.4",
    "json5==0.10.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2024.10.1",
    "mypy-extensions==1.1.0",
    "nodeenv==1.9.1",
    "numpy==2.3.2",
    "odfpy==1.4.1",
    "openpyxl==3.1.5",
    "packaging==24.2",
    "pandas==2.3.1",
    "pathspec==0.12.1",
    "pillow==11.3.0",
    "ffmpeg-python==0.2.0",
    "platformdirs==4.3.8",
    "polib==1.2.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic_core==2.33.2",
    "PyJWT==2.10.1",
    "PyMuPDF==1.26.3",
    "python-calamine==0.3.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyxlsb==1.0.10",
    "PyYAML==6.0.2",
    "referencing==0.35.1",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rpds-py==0.22.3",
    "s3transfer==0.6.2",
    "sgmllib3k==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tqdm==4.67.1",
    "typing_extensions==4.14.1",
    "tzdata==2024.2",
    "urllib3==1.26.20",
    "virtualenv==20.29.3",
    "whitenoise==6.9.0",
    "xlrd==2.0.2",
    "XlsxWriter==3.2.5",
    "markdown==3.8.2",
    "Pygments==2.19.2",
]

[tool.poetry.group.dev.dependencies]
    black = "25.1.0"
    pre_commit = "4.1.0"
    djlint = "1.36.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
